silver-tier/requirements.txt
watchdog
pytest
google-auth
google-auth-oauthlib
google-api-python-client
